Refactor api for centralize methods
